#214 feat(fe): React-Query 부분 연결 구현
